requirements.txt
Django~=4.1.7
fontawesomefree==6.3.0
djangorestframework~=3.14.0